refactoring

+++ ClientConnection/abstract_client_connection.h
@@ -10,7 +10,7 @@ class AbstractClientConnection : public QObject {
   virtual ~AbstractClientConnection();
 
   virtual void onInstanceThreadStarted(void) = 0;
-  virtual size_t getId(void) const = 0;
+  virtual int32_t getId(void) const = 0;
   virtual void reset(void) = 0;
 
  private:
